# docs/swagger.yaml
basePath: /
definitions:
  github_com_Ammar022_secure-ai-chat-backend_internal_chat_dto.SendMessageRequest:
    properties:
      question:
        description: |-
          Question is the user's prompt.  Required, 1–4000 characters.
          The validator rejects blank strings and anything exceeding the limit.
        example: What is quantum computing and how does it work?
        maxLength: 4000
        minLength: 1
        type: string
    required:
    - question
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody:
    properties:
      code:
        type: string
      details:
        description: optional field-level errors
      message:
        type: string
    type: object
...
  internal_chat_controller.messageResponse:
    properties:
      answer:
        example: This is a simulated AI response to your question.
        type: string
      completion_tokens:
        example: 85
        type: integer
      created_at:
        example: "2024-03-21T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      prompt_tokens:
        example: 12
        type: integer
      question:
        example: What is quantum computing and how does it work?
        type: string
      response_time_ms:
        example: 743
        type: integer
      subscription_id:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      total_tokens:
        example: 97
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Production-grade AI chat backend with Auth0 authentication, DDD, quota
    management, and subscription bundles.
  title: Secure AI Chat Backend
  version: "1.0"
paths:
...
  /api/v1/chat/messages:
    get:
      description: Returns a paginated list of the authenticated user's chat history,
        newest first.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page, 1–100 (default: 20)'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated messages
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_chat_controller.messageResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: List chat messages
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Submits a question, validates quota (free tier or subscription
        bundle), calls the mocked AI, stores and returns the result. Deducts from
        the subscription bundle with the most recent remaining quota.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567) — must be within
          ±5 min of server time
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID (rejected if reused —
          anti-replay)
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_chat_dto.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message stored and AI response returned
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_chat_controller.messageResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "402":
          description: No quota available — subscribe to continue
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Send a message to the AI
      tags:
      - chat
  /api/v1/chat/messages/{id}:
    get:
      description: Returns one chat message by ID. Users can only retrieve their own
        messages.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Message UUID (e.g. 550e8400-e29b-41d4-a716-446655440000)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chat message
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_chat_controller.messageResponse'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a single chat message
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: Auth0 JWT — prefix with "Bearer ". Obtain via Auth0 dashboard (APIs
      → Test tab) or your Auth0 login flow.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"